feat: Add method to create semantic view in Database class
Added a new method `_create_semantic_view` in the `Database` class to create a view with semantic renaming of the columns for enhanced readability. This method uses ollama to generate SQL code with semantic names for the columns of a specified table. The view can be created either as a regular SQL view or an in-memory DuckDB view based on the `duckdb_view` parameter. Logs the created view's SQL code and column descriptions.